alx_travel_app/requirement.txt
# Requirements for Django project (ALX Travel App)
django
djangorestframework
django-cors-headers
drf-yasg
celery
django-environ
mysqlclient
requests

# Note: RabbitMQ should be installed separately via system package manager
# For Ubuntu/Debian: sudo apt-get install rabbitmq-server
# For macOS: brew install rabbitmq
# For Windows: Download from https://www.rabbitmq.com/download.html

